Simplify Jackson config
